+++ test_terra_accounts/urls.py
@@ -18,5 +18,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
+    path('api/', include('terra_settings.urls')),
     path('api/', include('terra_accounts.urls')),
 ]